Moving the Schema from the controller file to the model file
